+++ demo/src/main/java/com/example/demo/cqrs/port/command/UserAddExpense.java
@@ -6,6 +6,6 @@ import org.jspecify.annotations.NonNull;
 
 public interface  UserAddExpense {
 
-    public abstract UUID createExpenseUser(@NonNull String email,int total_amount,String note,@NonNull UUID category_id,@NonNull int payment_status);
+    public abstract UUID createExpenseUser(@NonNull String email, double total_amount, String note, @NonNull UUID category_id, @NonNull int payment_status);
 
 }
